Thu Jan 16 06:48:32 UTC 2025
## Detailed Changes

### Functions

no changes

### Agencies

1 row changed

  citation: VA 4509
    _version_: "1816062777064685568" => "1821331040340279308"
    date_range: "["[1868 TO 1973]"]" => "["[1868 TO 1974]"]"
    description: "None" => "Tabilk Common School began operating in September 1868 in the town of Tabilk, Shire of Goulburn. It is likely to have originally operated out of a local house, church, or hall, although the exact building is not known. In August 1869, an acre of land was purchased for the school, and a residence built there the following year. Following naming convention set by the Education Act 1872, the name of the school was changed to Tabilk Common School.<br />
<br />
In 1893, the Public Works Department (VA 669) erected a new school building in place of the original. In 1914 a fire destroyed the teacher&rsquo;s residence, which was then replaced with a school building and teacher&rsquo;s quarters removed from the closed Wormangel State School School (No. 1446) the same year.<br />
<br />
As of December 1926, only 8 children were enrolled. The school continued to have low attendance in the following decades, though it had increased to 17 by 1970.<br />
<br />
In 1967, the school building was in bad repair and required extensive works. The Education Department (VA 714) agreed to only minor and urgent works carried out when needed, with the building due to be replaced with a portable when one became available. The following year, the teacher&rsquo;s residence was condemned and replaced with the schoolhouse from Goulburn Weir State School (School No. 3358) in 1970.<br />
<br />
In 1970, there was a statewide decision to change the use of the term &lsquo;State School&rsquo; to refer to all primary, secondary and technical schools instead of just primary schools, and for all primary schools to use &lsquo;Primary School&rsquo; in the name. Accordingly, the school name was likely changed to Tabilk Primary School around this time.<br />
<br />
From December 1973 the school was unstaffed, due to the small number of pupils in attendance. Though the large majority of residents opposed the decision, it was officially closed on the 24th of October 1974.<br />
<br />
Following its closure, the school building was moved to Seymour Primary School (School No. 547), while the residence and land were sold at public auction in 1984.<br />
&nbsp;"
    end_dt: "1973" => "1974"
    name.official_title.date_range: "[1868 TO 1973]" => "[1868 TO 1974]"
    name.official_title.end_dt: "1973" => "1974"
    name.official_title.title: "Tabilk (Primary School No. 951)" => "Tabilk Primary School (also known as Tabilk Common School (1868 – c. 1872), Tabilk State School (c. 1872 – 1970), School No. 951)"
    timestamp: "1718080811" => "1736912061"
    title: "Tabilk (Primary School No. 951)" => "Tabilk Primary School (also known as Tabilk Common School (1868 – c. 1872), Tabilk State School (c. 1872 – 1970), School No. 951)"

### Series

no changes

### Consignments

no changes