Thu Jan 30 06:48:52 UTC 2025
Functions:

Agencies: 1 row changed

  citation: VA 2984
    _version_: "1816043474273697793" => "1822599393220493312"
    description: "Establishment<br/><br/>The Transport (Amendment) Act 1989 (No.44) drew together all responsibilities relating to the provision of public transport services under a single agency, the Public Transport Corporation.  Prior to the establishment of the Corporation responsibility was shared by the State Transport Authority commonly known as V/Line (VA 1038) (primarily operating outside the Melbourne metropolitan area) and the Metropolitan Transit Authority (VA 1044) primarily operating within the metropolitan area.<br/><br/>Functions<br/><br/>The functions of the Public Transport Corporation, as laid down by its establishing legislation, are essentially an amalgamation of those carried out by the two agencies previously responsible for the provision of public transport services.  They include:<br/><br/>to provide, manage and operate transport services and facilities for passengers and freight<br/>to arrange with other transport operators, people and bodies for the provision of transport services<br/>to control and maintain all railway and railway ancillary equipment in the state<br/>to co-ordinate transport services<br/>to develop integrated ticket systems<br/>to develop new modes of transport<br/>to promote and undertake research into any matter related to the performance of its functions, powers or duties<br/>to provide, manage and operate tourist and recreational facilities<br/>to compete for work on the open market.<br/><br/>Administrative Arrangements<br/><br/>The Public Transport Corporation is subject to the control and general direction of the Director-General of Transport and, ultimately, the Minister for Transport.  A fourteen-member Public Transport Corporation Board is appointed by the Minister to provide advice to the Chief Executive of the Corporation.<br/><br/>The Public Transport Corporation is also assisted by a Workshop Board comprising nine or more members representing the Corporation and staff and management from the Corporation's railway workshops throughout the state.  This Board is responsible to the Corporation for the administration and management of the workshops.<br/><br/>Although the merger of the State Transport Authority (VA 1038) and the Metropolitan Transit Authority (VA 1044) to form the Public Transport Corporation brought responsibility for both metropolitan and country transport services within the ambit of the one agency, it is expected that the Corporation will continue to operate with two distinct functional divisions.  Corporate and management services are, however, being integrated and it is likely that a common record-keeping system will be established.  Further research will be necessary to determine the extent to which the functional division is retained in the Corporation's structure in the long term.<br/>" => "<em>Establishment</em><br />
<br />
The <em>Transport (Amendment) Act 1989 (No.44)</em> drew together all responsibilities relating to the provision of public transport services under a single agency, the Public Transport Corporation. Prior to the establishment of the Corporation responsibility was shared by the State Transport Authority commonly known as V/Line (VA 1038) (primarily operating outside the Melbourne metropolitan area) and the Metropolitan Transit Authority (VA 1044) primarily operating within the metropolitan area.<br />
<br />
<em>Functions</em><br />
<br />
The functions of the Public Transport Corporation, as laid down by its establishing legislation, are essentially an amalgamation of those carried out by the two agencies previously responsible for the provision of public transport services. They included:<br />
<br />
to provide, manage and operate transport services and facilities for passengers and freight<br />
to arrange with other transport operators, people and bodies for the provision of transport services<br />
to control and maintain all railway and railway ancillary equipment in the state<br />
to co-ordinate transport services<br />
to develop integrated ticket systems<br />
to develop new modes of transport<br />
to promote and undertake research into any matter related to the performance of its functions, powers or duties<br />
to provide, manage and operate tourist and recreational facilities<br />
to compete for work on the open market.<br />
<br />
<em>Administrative Arrangements</em><br />
<br />
The Public Transport Corporation was subject to the control and general direction of the Director-General of Transport and the Minister for Transport. A fourteen-member Public Transport Corporation Board was appointed by the Minister to provide advice to the Chief Executive of the Corporation.<br />
<br />
The Public Transport Corporation was also assisted by a Workshop Board comprising nine or more members representing the Corporation and staff and management from the Corporation&#39;s railway workshops throughout the state. This Board was responsible to the Corporation for the administration and management of the workshops.<br />
<br />
<em>Abolishment</em><br />
<br />
In 1996 the Government commenced the corporatisation of the Public Transport Corporation&rsquo;s business units, facilitated through amendments to the <em>Rail Corporations Act 1996</em>. In July 1997, the Government established the V/Line Freight Corporation as a separate corporatised entity with responsibility for the operation of the non-electrified Victorian intrastate rail freight business. Subsequently, in July 1998 property, rights, liabilities and staff of the Public Transport Corporation were further transferred to the remaining corporatised transport businesses, including the following:<br />
<br />
- Met Train 1 (trading as Bayside Trains) provided all metropolitan train services to and from Melbourne&rsquo;s northern, western and south-eastern suburbs through a network of 9 rail lines;<br />
<br />
- Met Train 2 (trading as Hillside Trains) provided all metropolitan train services to and from Melbourne&rsquo;s eastern and north-eastern suburbs through a network of 6 rail lines; &nbsp;<br />
<br />
- Met Tram 1 (trading as Swanston Trams) provided tram services on 18 routes operating from 4 depots to the north-western and south-eastern suburbs, and the Melbourne central business district;&nbsp;<br />
<br />
- Met Tram 2 (trading as Yarra Trams), provided tram services on 10 routes operating from 4 depots, encompassing both the north-eastern and southern suburbs as well as the central business district, and;&nbsp;<br />
<br />
- V/Line Passenger Corporation provided a network of integrated train and coach services to towns and cities across regional Victoria.&nbsp;"
    timestamp: "1614748017" => "1738129222"

Series:

Consignments: 1 row changed

  citation: VPRS 3426/P0009
    _version_: "1820787459838967808" => "1822599393199521794"
    date_range: "["[1859 TO 2006]"]" => "["[1859 TO 1954]"]"
    end_dt: "2006" => "1954"
    end_dt_qual: "?" => "None"
    timestamp: "1736316070" => "1738115393"