Fri Feb  7 06:49:39 UTC 2025
## Detailed Changes

### Functions

no changes

### Agencies

2 rows changed

  citation: VA 5300
    _version_: "1816063262299521025" => "1823324169781116929"
    description: "Alexandra Secondary College (School No. 7505) (VA 5300)&nbsp;&nbsp;first opened it&rsquo;s doors as Alexandra High School in 1953 on a shared site with Alexandra Higher Elementary School (School No. 912). A new site for a High School was purchased in 1945 and the High School relocated to a newly constructed facility in 1957. In 1978, a Technical component was established.<br />
<br />
Alexandra Secondary College is a co-education secondary school situated in the rural township of Alexandra in North-Eastern Victoria. Students are drawn from the greater district of Alexandra which includes the townships of Marysville, Narbelthong, Buxton, Taggerty, Thornton, Eildon, Molesworth, Yarck and Merton.&nbsp;<br />
<br />
In 1990 the school underwent a name change to Alexandra Secondary College following the Change of School Name and Reclassification of Schools Executive Memorandum No 314 from the Ministry of Education (VA 1112).&nbsp;" => "Alexandra Secondary College (School No. 7505) (VA 5300)&nbsp;&nbsp;first opened it&rsquo;s doors as Alexandra High School in 1953 on a shared site with Alexandra Higher Elementary School (School No. 912), absorbing the secondary classes of that school. A new site for a High School was purchased in 1945 and the High School relocated to a newly constructed facility in 1957. In 1978, a Technical component was established.<br />
<br />
Alexandra Secondary College is a co-education secondary school situated in the rural township of Alexandra in North-Eastern Victoria. Students are drawn from the greater district of Alexandra which includes the townships of Marysville, Narbelthong, Buxton, Taggerty, Thornton, Eildon, Molesworth, Yarck and Merton.&nbsp;<br />
<br />
In 1990 the school underwent a name change to Alexandra Secondary College following the Change of School Name and Reclassification of Schools Executive Memorandum No 314 from the Ministry of Education (VA 1112).&nbsp;"
    timestamp: "1730350598" => "1738795640"

  citation: VA 5444
    _version_: "1823233569703591936" => "1823324169789505537"
    timestamp: "1738650964" => "1738805345"

### Series

no changes

### Consignments

1 row changed

  citation: VPRS 3426/P0009
    _version_: "1823233572121608193" => "1823324169590276097"