Fri Apr  4 06:52:19 UTC 2025
Functions:

Agencies: 6 rows changed, 1 row added

6 rows changed

  citation: VA 553
    _version_: "1828308126597644288" => "1828398722733572096"
    timestamp: "1743546625" => "1743645537"

  citation: VA 559
    _version_: "1828308126624907264" => "1828398722737766400"
    timestamp: "1743547735" => "1743645564"

  citation: VA 593
    _version_: "1828217530784677888" => "1828398722740912128"
    timestamp: "1743487317" => "1743645597"

  citation: VA 4347
    _version_: "1828308126611275777" => "1828398722743009280"
    timestamp: "1743547327" => "1743645619"

  citation: VA 5597
    _version_: "1828308126615470080" => "1828398722745106432"
    timestamp: "1743547367" => "1743645638"

  citation: VA 5598
    _version_: "1828308126618615808" => "1828398722747203584"
    timestamp: "1743547411" => "1743645662"

1 row added

  _id: 147EB9AD-F3AA-11E9-AE98-A1F82E009A80
  _version_: 1828398722768175104
  category: Agency
  citation: VA 1685
  date_range: ["[1930 TO 2021]"]
  description: First opened in 1930, the Edenhope Bush Nursing Hospital was a community funded hospital managed by the Bush Nursing Association of Victoria. The original hospital consisted of 2 wards containing 5 beds each, as well as a hall, sitting room, maternity room, operating theatre, nurses quarters, kitchen, pantry, bathroom and laundry.<br />
<br />
In February 1946, the hospital was formally registered with the Charities Board of Victoria (VA 2707) to receive government funding. In December 1950 its name changed to Edenhope District Hospital, then the following year it&nbsp;was deregistered as a charity and became a fully-funded public hospital under the Hospital and Charities Commission. The Edenhope District Hospital Society was registered as a charity between December 1951 and February 1953 to manage funding until the hospital was incorporated under the <em>Hospital and Charities Act 1948</em> on 4 November 1953.<br />
&nbsp;<br />
In 1959 the hospital became an annex of Hamilton Base Hospital - known at the time as Glenelg Base Hospital (VA 1268) - along with Heywood and District Memorial Hospital (VA 1691), Macarthur and District Memorial Hospital (VA 1704), Penshurst and District Memorial Hospital (VA 1721) and later Willaura and District Hospital (VA 1749).&nbsp;Although the Hamilton Base Hospital provided administrative services to the annexed hospitals, the Edenhope Hospital continued to have its own President and Managing Committees. The Hamilton Base Hospital continued to administer the Edenhope Hospital until 1976, when it was formally renamed the Edenhope and District Memorial Hospital.<br />
<br />
There were several developments and extensions to the hospital throughout the second half of the 20th century. A new nurse&rsquo;s home was opened for the hospital in 1955,&nbsp;and there were several extensions to and remodelling of the main hospital buildings in the 1960s to 1980s. The Bade Wing Nursing Home was opened on 25 October 1991 followed by The Lakes Hostel in March 1997. The Lakes Hostel was extended further in 2001. In October 1997 the Elsie Bennet Community Centre, a new Adult Day Centre, opened in conjunction with the Lakes Hostel. The Edenhope Hospital Medical Clinic was opened in October 2012 to provide GP and various allied health services; however, operation of this clinic was leased to Rural Doctors in 2020.<br />
<br />
In 2021, Edenhope and District Memorial Hospital was merged with Wimmera Health Care Group (VA 5501), Stawell Regional Health (VA 1736) and Ballarat Health Services (VA 5500) to form Grampians Health (VA 5499). The Hospital continued to operate under the name Grampians Health Edenhope.
  end_dt: 2021
  end_dt_qual: None
  entity: Agent
  identifier.PID.id: 147EB9AD-F3AA-11E9-AE98-A1F82E009A80
  identifier.PROV_ACM.id: VA 1685
  jurisdictional_coverage: ["Victoria"]
  name.aka.date_range: None
  name.aka.end_dt: None
  name.aka.end_dt_qual: None
  name.aka.start_dt: None
  name.aka.title: None
  name.official_title.date_range: [1976 TO 2021]
  name.official_title.end_dt: 2021
  name.official_title.end_dt_qual: None
  name.official_title.start_dt: 1976
  name.official_title.start_dt_qual: None
  name.official_title.title: Edenhope and District Memorial Hospital
  start_dt: 1930
  start_dt_qual: None
  status: Published
  timestamp: 1743646094
  title: Edenhope and District Memorial Hospital (previously known as Edenhope Bush Nursing Hospital 1930 - 1950; Edenhope and District Hospital 1950 - 1976)

Series: 8 rows changed

  citation: VPRS 3373
    _version_: "1828308126623858688" => "1828398722749300736"
    timestamp: "1743547672" => "1743645692"

  citation: VPRS 20422
    _version_: "1828217530781532161" => "1828398722751397888"
    timestamp: "1743487166" => "1743645725"

  citation: VPRS 20423
    _version_: "1828308126589255680" => "1828398722753495040"
    timestamp: "1743487909" => "1743645756"

  citation: VPRS 20787
    _version_: "1828217530779435008" => "1828398722758737920"
    timestamp: "1743485710" => "1743645791"

  citation: VPRS 20788
    _version_: "1828217530788872192" => "1828398722760835072"
    timestamp: "1743487657" => "1743645924"

  citation: VPRS 20789
    _version_: "1828308126602887168" => "1828398722761883648"
    timestamp: "1743546801" => "1743645952"

  citation: VPRS 20790
    _version_: "1828308126606032897" => "1828398722765029376"
    timestamp: "1743547231" => "1743645956"

  citation: VPRS 20791
    _version_: "1828308126621761536" => "1828398722766077952"
    timestamp: "1743547539" => "1743645991"

Consignments: