Wed May  7 06:54:44 UTC 2025
## Detailed Changes

### Functions

no changes

### Agencies

1 row changed

  citation: VA 3548
    _version_: "1827764545241743362" => "1831392197771001856"
    description: "<br />
Originally named &lsquo;Dana Street&rsquo;, this school was the first National School established in West Ballarat. &nbsp;The school&rsquo;s original bluestone building was built on Dana Street in 1856, and opened in 1857 with 50 pupils in attendance. By the end of the year this number had more than doubled.<br />
<br />
The first teachers at the school were Charles Martin and his wife; during this time the school was colloquially known as &ldquo;Mr Martin&rsquo;s School&rdquo;.<br />
<br />
In 1862 it became a Common School, and in 1876, a State School. A new, larger brick building with fireplaces in each room was built the same year to house the then 1,100 pupils enrolled. An influx of students came from four local church schools that had closed due to lack of funding.<br />
<br />
The original bluestone building remained at the school and held evening classes until 1907, when it was used by Ballarat High School until 1910. The Ballarat Junior Technical School then used the building while their own new school was built.<br />
<br />
In 1927 the Ballarat Teachers College moved into the bluestone building for one year while the Ballarat East Town Hall was renovated. The College closed in 1931 due to the Great Depression, but re-opened in the bluestone building in 1946 and stayed there until 1958.<br />
<br />
Sometime in 1970 the school was renamed Ballarat Primary School, also known as Dana Street Primary School. In 1977 Jubilee gates were built at the school for the 25th anniversary of the reign of Queen Elizabeth II.<br />
<br />
Several major building projects were completed in the 1990s. The school was remodelled, and the old bluestone building was converted into a multipurpose room. In 2012 it once again became a classroom.<br />
<br />
As of 2025, the school is still in operation with an average yearly enrolment of 250 students." => "Originally named &lsquo;Dana Street&rsquo;, this school was the first National School established in West Ballarat. The school&rsquo;s original bluestone building was built on Dana Street in 1856, and opened in 1857 with 50 pupils in attendance. By the end of the year this number had more than doubled.<br />
<br />
The first teachers at the school were Charles Martin and his wife; during this time the school was colloquially known as &ldquo;Mr Martin&rsquo;s School&rdquo;.<br />
<br />
In 1862 it became a Common School, and in 1876, a State School. A new, larger brick building with fireplaces in each room was built the same year to house the then 1,100 pupils enrolled. An influx of students came from four local church schools that had closed due to lack of funding.<br />
<br />
The original bluestone building remained at the school and held evening classes until 1907, when it was used by Ballarat High School until 1910. The Ballarat Junior Technical School then used the building while their own new school was built.<br />
<br />
In 1927 the Ballarat Teachers College moved into the bluestone building for one year while the Ballarat East Town Hall was renovated. The College closed in 1931 due to the Great Depression, but re-opened in the bluestone building in 1946 and stayed there until 1958.<br />
<br />
Sometime in 1970 the school was renamed Ballarat Primary School, also known as Dana Street Primary School. In 1977 Jubilee gates were built at the school for the 25th anniversary of the reign of Queen Elizabeth II.<br />
<br />
Several major building projects were completed in the 1990s. The school was remodelled, and the old bluestone building was converted into a multipurpose room. In 2012 it once again became a classroom.<br />
<br />
As of 2025, the school is still in operation with an average yearly enrolment of 250 students."
    name.official_title.title: "Ballarat Primary School (Dana Street) (also known as Dana Street Primary School, previously known as Dana Street National School (1857 – 1862), Dana Street Common School (1862 – 1876), Dana Street State School (1876 - 1970), School No. 33" => "Ballarat Primary School (Dana Street) (School No. 33; also known as Dana Street Primary School; previously known as Dana Street National School 1857 – 1862; Dana Street Common School 1862 – 1876; Dana Street State School 1876 - 1970)"
    timestamp: "1743053023" => "1746497548"
    title: "Ballarat Primary School (Dana Street) (School No. 33, also known as Dana Street Primary School, previously known as Dana Street National School (1857 – 1862), Dana Street Common School (1862 – 1876), Dana Street State School (1876 - 1970)" => "Ballarat Primary School (Dana Street) (School No. 33; also known as Dana Street Primary School; previously known as Dana Street National School 1857 – 1862; Dana Street Common School 1862 – 1876; Dana Street State School 1876 - 1970)"

### Series

no changes

### Consignments

no changes