Improve git-scraping commit messages
When using the track script to do git-scraping to track the changes
in records available by the API improve the commit messages by
building on csv-diff to create new sections to show New Additions
and Openings.

As part of this I felt it was worthwhile to re-process all the
previous commits to be able to access this summarised information
easily. As noted in the README this does have some drawbacks but
I think on balance they are worth it.

I have kept the scripts to do this in place in case they are
useful in the future.