Wed Sep 10 06:52:46 UTC 2025
## Detailed Changes

### Functions

no changes

### Agencies

2 rows changed

  citation: VA 4710
    _version_: "1839092941540294656" => "1842807415182983169"
    description: "Under the provisions of the Guardianship and Administration Board Act 1986 No 58, two new agencies were created namely the Guardianship and Administration Board (VA 3027) and The Office of the Public Advocate.<br />
<br />
The Guardianship and Administration Board Act 1986 was passed in response to the Report of the Minister&#39;s Committee on Rights and Protective Legislation for intellectually handicapped persons, commonly known as the Cock&#39;s Committee Report. The report was originally commissioned by the Minister for Health (the Hon. W.A. Borthwick) in December 1980. Terms of reference were expanded to include other disabilities in September 1982. The final report was submitted to the Minister of Health (the Hon. T. Roper) in December 1982.<br />
<br />
The Act was proclaimed in a number of stages. In July 1986 the Public Advocate was appointed under Schedule 3 of the Act. In April 1987, the President of the Board was appointed and operational provisions relating to The Office of the Public Advocate were proclaimed. The Board was the Tribunal established to legally protect the personal and financial affairs of adult disabled persons including matters relating to medical and financial powers of attorney. &#39;Disability&#39; is defined under Section 3 of the Act as including intellectual impairment, mental illness, brain damage, physical disability and senility.<br />
<br />
The role of The Office of the Public Advocate was to advocate, promote and support the rights and interests of disabled persons. It frequently represented persons in Board hearings.<br />
<br />
Despite the interrelated role and a shared location with the Board, The Office of the Public Advocate was a separately constituted, administered and staffed agency." => "The role of the Office of the Public Advocate (OPA) is to advocate, promote and support the rights and interests of disabled persons. It frequently represented persons in hearings of the Guardianship Administration Board (VA 3027) from 1987 to 1998 and the Victorian Civil and Administration Tribunal (VCAT, VA 4146) from 1998.<br />
<br />
OPA provides advocacy, promotion and support for the rights and interests of disabled persons in accordance with the <em>Guardianship and Administration Act </em>1986 (no. 58). The Public Advocate acts as guardian of last resort. Under the provisions of the Act, two new agencies were created: namely the Guardianship and Administration Board (VA 3027) and The Office of the Public Advocate.<br />
<br />
The Act&nbsp;was passed in response to the Report of the Minister&#39;s Committee on Rights and Protective Legislation for Intellectually Handicapped persons, commonly known as the Cock&#39;s Committee Report. The report was originally commissioned by the Minister for Health (the Hon. W.A. Borthwick) in December 1980. Terms of reference were expanded to include other disabilities in September 1982. The final report was submitted to the Minister of Health (the Hon. T. Roper) in December 1982.<br />
<br />
The Act was proclaimed in a number of stages. In July 1986 the Public Advocate was appointed under Schedule 3 of the Act. In April 1987, the President of the Board was appointed and operational provisions relating to The Office of the Public Advocate were proclaimed. The Board was the Tribunal established to legally protect the personal and financial affairs of adult disabled persons including matters relating to medical and financial powers of attorney. &#39;Disability&#39; was defined under Section 3 of the Act as including intellectual impairment, mental illness, brain damage, physical disability and senility."
    name.official_title.start_dt_qual: "By" => "None"
    start_dt_qual: "By" => "None"
    timestamp: "1753857819" => "1757397951"

  citation: VA 5021
    _version_: "1839092942017396736" => "1842807415174594580"
    timestamp: "1753857893" => "1757393823"

### Series

no changes

### Consignments

no changes