Wed Oct 15 06:53:59 UTC 2025
## Detailed Changes

### Functions

no changes

### Agencies

2 rows changed

  citation: VA 3887
    _version_: "1839092942280589313" => "1845974537985851393"
    description: "None" => "Located in the Murray Valley area of Northern Victoria, this school officially opened as Baulkamaugh West State School on 28 January 1890. The name changed to Waaia State School (No. 2986) (VA 3887) by 1900. Another school then also known as &lsquo;Waaia&rsquo; became Waaia South State School (School No. 1950). The construction of the school building is uncertain, however it was likely a small one-room timber schoolhouse.<br />
<br />
On 12 March 1946, the school closed and pupils were conveyed to Numurkah (School No. 2134), but it re-opened on 6 June 1946. By 1952 the school was overcrowded, with an enrolment of 47 and students sitting three per desk. An unused school building from the closed Picola West school (School No. 2558) as well as outbuildings and a shelter pavilion were moved to Waaia the following year. By 1955 enrolment was 67 and expected to rise, and two timber classrooms were constructed on an additional site purchased for the school. Both the original and new sites were then in use, providing four classrooms total.<br />
<br />
In 1962 the two buildings on the original site were in poor repair and condemned. They were demolished the following year, and an additional building was added to the new site. Enrolments fluctuated between 50-90 over the following decade. The name of the school likely changed to Waaia Primary School around 1970, following a state-wide change to naming convention.<br />
<br />
In 1993 the school merged with Yalca South (School No. 2198) (VA 3894) to become Waaia Yalca South Primary School (No. 5401). The school was consolidated on the Waaia Primary School site and the Yalca South site was closed and later sold."
    timestamp: "1753857930" => "1760414615"
    title: "Waaia (Primary School No.2986; previously known as Baulkamaugh West 1890-?1892)" => "Waaia Primary School (School No. 2986; previously known as Baulkamaugh West 1890-c.1892; Waaia State School c.1892-c.1970)"

  citation: VA 5378
    _version_: "1839092942907637760" => "1845974537981657088"
    timestamp: "1753857990" => "1760412496"
    title: "Agnes State School (School No. 3043, previously Agnes River State School (1890 – c. 1950s)" => "Agnes State School (School No. 3043; previously Agnes River State School 1890 – c. 1950s)"

### Series

no changes

### Consignments

no changes